Conversion of the GMaps Geocoding functions to Nominatim compatible
This basically makes a "PokeAlarm OSM Edition"

TL;DR: Replace your `GMaps.py` with this one and read the configuration section below

Details:
- Geocoding and Reverse Geocoding are both working using a Nominatim server.
- Distances are not supported by Nominatim therefore any calls to that function will simply return the current default `unknown`.
- The Google API is no longer called for anything.

Notes:
Priority #1 was to make PA compatible with Nominatim without compromising the stability of the original branch so only what was absolutely necessary was modified (Gmaps.py). There was no attempt to cleanup of the code to get rid of the Google nomenclature, variables, distance components or internal checks that are no longer necessary. For that reason, this is to be considered a working drop-in replacement of a Frankenstein modification!

Installation:
`GMaps.py` was last officially modified in March 2018 so I would simply replace your existing `\PokeAlarm\LocationServices\GMaps.py` with the one from this branch. Unless your branch decides to modify the calls to the functions inside that file, it should work indefinitely. You can also git clone as per usual but don’t expect this branch to be kept updated unless this specific file needs an update.

Configuration:
- Must have `gmaps-key: SomeRandomStringHere` in your `config.ini` and uncommented
- Must have `gmaps-rev-geocode: yes` in your `config.ini` and uncommented
- Might have to edit `GMaps.py` to set your Nominatim server address if it's different from the default: http://127.0.0.1:7070/
- Done and enjoy life Google Free!